# Performance tuning: First batch