This PR dramatically improves performance in certain situations involving Indexed drawing that requires vertex format conversion.

On one of my computers, this made the Dashboard dump from ~5fps to 40fps
On my other computer, the dashboard jumped from 20fps to a perfect 60fps.

Many titles should show at least some improvement, but others show no improvement at all.
There is no compatibiltiy impact for these changes, it's a completely free optimisation

There's still plenty of work that could be done, but this is a nice, very low risk set of changes. More will follow in the coming days/weeks.