@ESPPIRAL 2018-08-14
Otogi went from lower than 30 fps to mostly 60 fps

![bandicam 2018-08-14 23-40-08-159](https://user-images.githubusercontent.com/29869781/44120113-954386aa-a01b-11e8-9619-4eac03164d6d.png)