@ESPPIRAL 2018-08-15
The dashboard runs more fluid now too.

![bandicam 2018-08-15 11-05-06-960](https://user-images.githubusercontent.com/29869781/44140717-324a47b8-a07b-11e8-9018-e685d0c6c5e3.png)